# openapi.yaml
openapi: 3.0.3
info:
  title: TSTS - Technical Support Ticketing System API
  description: |
    Comprehensive API documentation for the Technical Support Ticketing System (TSTS).
    
    ## Overview
    TSTS is a full-featured ticketing system designed for university IT support operations.
    It includes ticket management, chat, knowledge base, AI assistance, and comprehensive reporting.
    
    ## Authentication
    The API supports multiple authentication methods:
    - **Bearer Token**: JWT-based authentication for web and mobile clients
    - **API Keys**: For third-party integrations with scope-based access control
    - **WebAuthn**: Passwordless authentication using trusted devices
    
    ## Base URLs
    - Production: `https://api.tsts.local`
    - Development: `http://localhost:3000`
    
    ## Rate Limiting
    - 1000 requests per 15 minutes per IP address
    - Rate limit headers included in responses
    
    ## Pagination
    Most list endpoints support pagination with `page` and `limit` query parameters:
    - Default page: 1
    - Default limit: 20
    - Maximum limit: 200
    
    ## Internationalization
    The API supports English and Arabic. Many fields use JSONB format:
    ```json
    {
      "name": {
        "en": "Computer Science",
        "ar": "علوم الحاسب"
      }
    }
    ```
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "message": "Error description",
      "statusCode": 400,
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party integrations
  schemas:
...
    CustomForm:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: [text, email, textarea, select, checkbox, radio, date, number]
              label:
                type: string
              required:
                type: boolean
              options:
                type: array
                items:
                  type: string
        token:
          type: string
        isGlobal:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        keyPrefix:
          type: string
          example: "tsts_abc123"
        zones:
          type: array
          items:
            type: string
          example: ["tickets", "knowledge_base"]
        methods:
          type: array
          items:
            type: string
          example: ["GET", "POST"]
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v1/api-integrations/keys:
    get:
      summary: List API keys
      tags: [API Integrations]
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'

    post:
      summary: Create API key
      tags: [API Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, zones, methods]
              properties:
                name:
                  type: string
                description:
                  type: string
                zones:
                  type: array
                  items:
                    type: string
                methods:
                  type: array
                  items:
                    type: string
                expiresAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: Full API key (shown only once)
                  apiKey:
                    $ref: '#/components/schemas/ApiKey'
  /api/v1/api-integrations/keys/{id}:
    patch:
      summary: Update API key
      tags: [API Integrations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                zones:
                  type: array
                  items:
                    type: string
                methods:
                  type: array
                  items:
                    type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: API key updated

    delete:
      summary: Revoke API key
      tags: [API Integrations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key revoked

  # Users Endpoints
...
  /api/v1/custom-forms:
    get:
      summary: List custom forms
      tags: [Custom Forms]
      responses:
        '200':
          description: List of custom forms
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomForm'

    post:
      summary: Create custom form
      tags: [Custom Forms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, fields]
              properties:
                title:
                  type: string
                description:
                  type: string
                fields:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Custom form created
  /api/v1/custom-forms/public/{token}:
    get:
      summary: Get public form by token
      tags: [Custom Forms]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public form details

  # Reports Endpoints
...